learning/pyproject.toml
[project]
name = "learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.49",
    "langchain-groq>=0.2.1",
    "langchain-openai>=0.2.12",
    "langchain-pinecone>=0.2.0",
    "langchain>=0.3.22",
    "langsmith>=0.2.2",
    "pinecone>=5.3.1",
    "python-dotenv>=1.0.1",
    "weave>=0.51.23",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai==2.0.11",
    "langchain-experimental>=0.3.4",
    "langgraph>=0.2.60",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "weaviate-client>=4.10.2",
    "langchain-weaviate>=0.0.3",
    "scikit-learn>=1.6.0",
    "duckduckgo-search>=7.2.0",
    "langchain-ollama>=0.2.2",
    "langchain-ai21>=1.0.1",
    "sseclient-py>=1.8.0",
    "langchain-mongodb>=0.4.0",
    "langchain-chroma>=0.1.2",
    "lark>=1.2.2",
    "opik>=1.4.5",
    "autoflake>=2.3.1",
    "flake8>=7.1.1",
    "black>=25.1.0",
    "mypy>=1.14.1",
    "isort>=6.0.0",
    "tavily-python>=0.5.1",
    "email-validator>=2.2.0",
    "langchain-linkup>=0.1.5",
    "exa-py>=1.9.0",
    "langchain-exa>=0.2.1",
    "newsapi-python>=0.2.7",
    "colorlog>=6.9.0",
    "chainlit>=2.5.5",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "loguru>=0.7.3",
]
